// streaming_client/src/lib/types.ts
export type Anime = {
	id: IDWrapper;
	mal: string;
	shinden: string | null;
	discord_role_id: string | null;
	title: string;
	description: string | null;
	anime_type: string;
	episodes: number;
	status: string;
	in_progress: boolean;
	release_time: string;
	aired: string;
	source: string;
	tags: TagOrStudio[];
	studio: TagOrStudio[];
	image: string;
	date: string;
	slug: string;
	alternative_title: string;
};

export type GenericResponse = string | Anime | UserModelSuperUser;

export interface RResponse {
	status: number;
	message: GenericResponse;
}

export type AnimeList = Anime[];

export type MultiSelectType = {
	value: string;
	name: string;
	color: string;
};

export type ComboBoxType = {
	value: string;
	label: string;
};

export type UserResponse = {
	status: number;
	message: UserModelSuperUser;
};

export type UserModelSuperUser = {
	id: IDWrapper;
	name: string;
	password: string;
	avatar: string;
	email: string;
	description: string;
	translator: boolean;
	proofreader: boolean;
	uploader: boolean;
	editor: boolean;
	published: boolean;
	created_at: string;
	updated_at: string;
	is_superuser: boolean;
	is_staff: boolean;
	is_active: boolean;
	is_banned: boolean;
	roles: Role[] | null;
	slug: string;
};